+++ CHANGELOG.md
@@ -57,6 +57,7 @@ Copyright (C) 2025 - Mr. Lima
 - Added support for `list.default` that runs by default on all List choices
 - `query` command no longer consuming until connection end
 - Added button **Refresh** in _Process Viewer_
+- Command `start` can start services installed on `/etc/services`
 
 ### 📦 Packages
 - **JBuntu** updated to _1.4_